# [2/3]  Optimize Slime Update Weights: Avoid GPU-to-CPU Device Sync when update expert weights