<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Motivation

- 1/3: https://github.com/sgl-project/sglang/pull/8751 （23s => 22s)
- 2/3: https://github.com/sgl-project/sglang/pull/8753  (22s => 12s)
- 3/3: https://github.com/sgl-project/sglang/pull/8754  (12s  => 10s)


- This PR[2/3]  Optimize Slime Update Weights: Avoid GPU-to-CPU Device Sync when update expert weights

Below graph shows a unnecessary device sync from Device to Host
<img width="727" height="281" alt="Screenshot 2025-08-04 at 2 05 31 PM" src="https://github.com/user-attachments/assets/066ee508-3671-4c7f-b5ac-f05f47157f04" />

This PR maps to optimization 2
<img width="776" height="599" alt="Screenshot 2025-08-03 at 11 44 18 PM" src="https://github.com/user-attachments/assets/c1990ff5-4c58-4394-b53f-7ae8dc45244c" />


<!-- Explain the purpose of this PR and the goals it aims to achieve. -->

## Modifications

<!-- Describe the changes made in this PR. -->

## Accuracy Test

<!-- If this PR affects model-side code (e.g., kernels, model architecture), please provide accuracy test results. Ref: https://docs.sglang.ai/references/accuracy_evaluation.html -->

## Benchmark & Profiling

<!-- If this PR is expected to impact performance, please provide benchmark and profiling results. Ref: https://docs.sglang.ai/references/benchmark_and_profiling.html -->

## Checklist

- [ ] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [ ] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [ ] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [ ] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [ ] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.
- [ ] Please feel free to join our Slack channel at https://slack.sglang.ai to discuss your PR.